Cleanup Recipe class
Make the interface more clear. The update method updates the
checksum of the recipe which is clear by the name of the
method it calls: writeRecipeChecksum.

There should be no connection between a recipe and a job, but
a recipe can check if something has changed in its configuration
or metadata. So there is a method uptodate? which checks that