Added Uri class to control access
Simple uri's like local paths or remote locations via http
doesn't require a pre processing. However there are more
complex uri's like the path to an iso which requires to
loop mount the device prior to use it. It's also required to
cleanup the state of such devices properly. It also makes
sense to ask an Uri class if the resource is remote or
local and maybe more useful messages can be passed to an
Uri object in the future