Fixed handling of relative repo source paths
Relative source paths can be set relative to the image
description. Inside of the container the image description
is mapped below /vagrant. If the relative path is leaving
the image description directory it is in the user's
responsibility to map the repo to the right place inside
of the container so that the relative path points correctly
to the repository. This fixes #7