Refactor job command invokation
Make sure the basic access to the buildsystem is provided by
the buildsystem factory and not created as part of the job.
This allows to run job commands within a buildservice by the
nature of the buildservice and not hard coded to the use
of ssh. This change has been in preparation to add a native
docker build service class