Fixed calculateDigest
The calculation of the recipe digest is based on the data which
build the image description. However if people use a repository
as part of the image description it should not be taken into
consideration, because a solver operation will find potential
repository changes. Therefore the list of files and directories
belonging to an image description is limited to the top level
entries and the recursive search for the optioal overlay
root directory