Make sure host system /dev is connected to guest
docker does not run a full devtmpfs mount inside of the
guest. It is only a tmpfs. This leads to the problem
if commands are supposed to create device nodes below
/dev that they never appear in the container. Therefore
the /dev location is bind mounted to the host in order
to allow for image building. The security implications
are well known, but there is currently no other choice
and the container already runs as a privileged
container. Fixes bnc#1013619, Fixes #12